信息工程学院大屏幕展示/requirements.txt
# 招聘数据分析系统依赖包

# 数据处理
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0

# 网络请求
requests>=2.25.0
bs4>=0.0.1
fake-useragent>=0.1.11

# 文本处理
jieba>=0.42.1
wordcloud>=1.8.1

# 数据可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# 定时任务
schedule>=1.1.0

# 环境变量管理
python-dotenv>=0.19.0

# Web应用
Flask>=2.0.0